Give padding to main instead of body
Earlier I added left and right padding to the body element in order
to keep the main content narrow on medium size screen.

Now I moved that padding to the main content instead,
so the header with the logo remains on the left side of the screen
all the time.